gitlab: Delete existing package files before upload
Implement a deletion mechanism for existing package files in the GitLab
generic package registry before uploading new ones. This change prevents
the creation of duplicate package files when a file with the same name
and version already exists in the registry.

When a package file is to be uploaded, the `delete_existing_files`
function now queries the GitLab API to identify and remove any files
that match the target package name, package version, and filename. This
process ensures that the registry retains only the latest version of a
given file, streamlining package management and reducing clutter.

Signed-off-by: Javier Tia <javier.tia@linaro.org>